"Updated constants.h, libOLED.cpp, and utilsfunc.h: changed BUFFER_SIZE type to uint16_t, removed unnecessary code in draw_line function, added _SIGN function, and made __writter function inline."

+++ constants.h
@@ -37,7 +37,7 @@ enum DISPLAY_COMMANDS : uint8_t{
 
 };
 
-constexpr uint8_t  BUFFER_SIZE = 1024;  /// MAX_X * MAX_Y / 8
+constexpr uint16_t BUFFER_SIZE = 1024;  /// MAX_X * MAX_Y / 8
 constexpr uint16_t EEPROM_SIZE = 1024;
 constexpr uint16_t EEPORM_PAGE_SIZE = 4;
 constexpr uint16_t FLASH_SIZE = 30720;
